CLIM-1367 (2): Atom1 It is simply Math.round, do we really want to move things around?
